Support editing of any numeric type swift with either a UIStepper, or now also a UISlider.
Any tweak can determine its own default editing mode on init.

If no default is supplied upon init, a tweak will use the SwiftTweaks.defaultNumericEditingStyle global.  The default value for this global is NumericEditStyle.stepper.  This global can be updated at runtime, if desired.

On iOS 11+, both the floating and collection view controllers support a trailing swipe action that allows changing the current editing mode for that tweak.

The slider logic supports the same min, max, and stepSize properties that the stepper does.